refactor: reorganiza diretórios do projeto
