+++ src/domain/customer/repository/customer.repository.interface.ts
@@ -1,0 +1,6 @@
+import { Customer } from "../entity/customer";
+import { IRepository } from "../../@shared/repository/repository.interface";
+
+export interface ICustomerRepository extends IRepository<Customer> {
+
+}
