--- src/domain/event/product/handler/send-email-when-product-is-created.handler.ts
@@ -1,9 +1,0 @@
-import { IEventHandler } from "../../@shared/event-handler.interface";
-import { ProductCreatedEvent } from "../product-created.event";
-
-export class SendEmailWhenProductIsCreateHandler implements IEventHandler<ProductCreatedEvent> {
-    handle(event: ProductCreatedEvent): void {
-        console.log(`Sending e-mail to ${JSON.stringify(event.data)}`);
-    }
-
-}
