+++ src/domain/order/repository/order.repository.interface.ts
@@ -1,0 +1,6 @@
+import { IRepository } from '../../@shared/repository/repository.interface';
+import { Order } from '../entity/order';
+
+export interface IOrderRepository extends IRepository<Order> {
+
+}
