+++ src/domain/product/repository/product.repository.interface.ts
@@ -1,0 +1,6 @@
+import { Product } from "../entity/product";
+import { IRepository } from "../../@shared/repository/repository.interface";
+
+export interface IProductRepository extends IRepository<Product> {
+
+}
