--- src/domain/repository/customer.repository.interface.ts
@@ -1,6 +1,0 @@
-import { Customer } from "../entity/customer";
-import { IRepository } from "./repository.interface";
-
-export interface ICustomerRepository extends IRepository<Customer> {
-
-}
