--- src/domain/repository/order.repository.interface.ts
@@ -1,6 +1,0 @@
-import { Order } from "../entity/order";
-import { IRepository } from "./repository.interface";
-
-export interface IOrderRepository extends IRepository<Order> {
-
-}
