--- src/domain/repository/product.repository.interface.ts
@@ -1,6 +1,0 @@
-import { Product } from "../entity/product";
-import { IRepository } from "./repository.interface";
-
-export interface IProductRepository extends IRepository<Product> {
-
-}
